replaced Integer with BigInteger

+++ src/main/java/org/example/Fibonacci.java
@@ -1,10 +1,11 @@
 package org.example;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface Fibonacci {
 
     void printFibonacci(int size);
 
-    List<Integer> getFibonacci(int size);
+    List<BigInteger> getFibonacci(int size);
 }
